Fix GetSubTree depth argument type
GetSubTree takes an integer for the depth, and not a size_t.
The upcoming mapper implementation needs it correct.

Tested:  No longer throws an exception with the fix.

+++ dbus.hpp
@@ -64,7 +64,7 @@ ObjectValueTree getManagedObjects(sdbusplus::bus::bus& bus,
  *                       services and interfaces.
  */
 DbusSubtree getSubtree(sdbusplus::bus::bus& bus, const std::string& root,
-                       size_t depth, const std::string& interface);
+                       int depth, const std::string& interface);
 
 /**
  * Get the D-Bus service name for the object path and interface from
